pyproject.toml
[tool.poetry]
name = "adeeny"
version = "0.1.0"
description = ""
authors = ["coderhanzo <reinerbielert@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
django-phonenumber-field = "^8.0.0"
mysqlclient = "^2.2.4"
djoser = "^2.2.3"
phonenumbers = "^8.13.40"
pillow = "^10.4.0"
djangorestframework-gis = "^1.0"
drf-extra-fields = "^3.7.0"
mysql-connector-python = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
